Reorganize end to end models tests
