+++ tests/e2e/models/libs/lib_unittest.yml
@@ -1,0 +1,182 @@
+# Copyright (c) 2024, RTE (https://www.rte-france.com)
+#
+# See AUTHORS.txt
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+# SPDX-License-Identifier: MPL-2.0
+#
+# This file is part of the Antares project.
+library:
+  id: basic
+  description: Basic library
+
+  port-types:
+    - id: flow
+      description: A port which transfers power flow
+      fields:
+        - id: flow
+
+  models:
+
+    - id: generator
+      description: A basic generator model
+      parameters:
+        - id: cost
+          time-dependent: false
+          scenario-dependent: false
+        - id: p_max
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: generation
+          lower-bound: 0
+          upper-bound: p_max
+      ports:
+        - id: injection_port
+          type: flow
+      port-field-definitions:
+        - port: injection_port
+          field: flow
+          definition: generation
+      objective: expec(sum(cost * generation))
+
+    - id: node
+      description: A basic balancing node model
+      ports:
+        - id: injection_port
+          type: flow
+      binding-constraints:
+        - id: balance
+          expression:  sum_connections(injection_port.flow) = 0
+    - id: spillage
+      description: A basic spillage model
+      parameters:
+        - id: cost
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: spillage
+          lower-bound: 0
+      ports:
+        - id: injection_port
+          type: flow
+      port-field-definitions:
+        - port: injection_port
+          field: flow
+          definition: -spillage
+    - id: unsupplied
+      description: A basic unsupplied model
+      parameters:
+        - id: cost
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: unsupplied_energy
+          lower-bound: 0
+      ports:
+        - id: injection_port
+          type: flow
+      port-field-definitions:
+        - port: injection_port
+          field: flow
+          definition: unsupplied_energy
+
+    - id: demand
+      description: A basic fixed demand model
+      parameters:
+        - id: demand
+          time-dependent: true
+          scenario-dependent: true
+      ports:
+        - id: injection_port
+          type: flow
+      port-field-definitions:
+        - port: injection_port
+          field: flow
+          definition: -demand
+
+    - id: short-term-storage
+      description: A short term storage
+      parameters:
+        - id: efficiency
+        - id: level_min
+        - id: level_max
+        - id: p_max_withdrawal
+        - id: p_max_injection
+        - id: inflows
+      variables:
+        - id: injection
+          lower-bound: 0
+          upper-bound: p_max_injection
+        - id: withdrawal
+          lower-bound: 0
+          upper-bound: p_max_withdrawal
+        - id: level
+          lower-bound: level_min
+          upper-bound: level_max
+      ports:
+        - id: injection_port
+          type: flow
+      port-field-definitions:
+        - port: injection_port
+          field: flow
+          definition: injection - withdrawal
+      constraints:
+        - id: Level equation
+          expression: level[t] - level[t-1] - efficiency * injection + withdrawal = inflows
+
+    - id: thermal-cluster-dhd
+      description: DHD model for thermal cluster
+      parameters:
+        - id: cost
+        - id: p_min
+        - id: p_max
+        - id: d_min_up
+        - id: d_min_down
+        - id: nb_units_max
+        - id: nb_failures
+          time-dependent: true
+          scenario-dependent: true
+      variables:
+        - id: generation
+          lower-bound: 0
+          upper-bound: nb_units_max * p_max
+          time-dependent: true
+          scenario-dependent: true
+        - id: nb_on
+          lower-bound: 0
+          upper-bound: nb_units_max
+          time-dependent: true
+          scenario-dependent: false
+        - id: nb_stop
+          lower-bound: 0
+          upper-bound: nb_units_max
+          time-dependent: true
+          scenario-dependent: false
+        - id: nb_start
+          lower-bound: 0
+          upper-bound: nb_units_max
+          time-dependent: true
+          scenario-dependent: false
+      ports:
+        - id: injection_port
+          type: flow
+      port-field-definitions:
+        - port: injection_port
+          field: flow
+          definition: generation
+      constraints:
+        - id: Max generation
+          expression: generation <= nb_on * p_max
+        - id: Min generation
+          expression: generation >= nb_on * p_min
+        - id: Number of units variation
+          expression: nb_on = nb_on[t-1] + nb_start - nb_stop
+        - id: Min up time
+          expression: sum(t-d_min_up + 1 .. t, nb_start) <= nb_on
+        - id: Min down time
+          expression: sum(t-d_min_down + 1 .. t, nb_stop) <= nb_units_max[t-d_min_down] - nb_on
+      objective: expec(sum(cost * generation))
