+++ tests/e2e/models/libs/standard_sc.yml
@@ -1,0 +1,210 @@
+# Copyright (c) 2024, RTE (https://www.rte-france.com)
+#
+# See AUTHORS.txt
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+# SPDX-License-Identifier: MPL-2.0
+#
+# This file is part of the Antares project.
+library:
+  id: basic
+  description: Basic library
+
+  port-types:
+    - id: flow
+      description: A port which transfers power flow
+      fields:
+        - id: flow
+    - id: emission
+      description: A port which transfers co2 emission
+      fields:
+        - id: emission
+
+  models:
+    - id: convertor
+      description: A basic convertor model
+      parameters:
+        - id: alpha
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: input
+          lower-bound: 0
+      ports:
+        - id: input_port
+          type: flow
+        - id: output_port
+          type: flow
+      port-field-definitions:
+        - port: input_port
+          field: flow
+          definition: -input
+        - port: output_port
+          field: flow
+          definition: input * alpha
+
+    - id: two_input_convertor
+      description: Two input convertor model
+      parameters:
+        - id: alpha1
+          time-dependent: false
+          scenario-dependent: false
+        - id: alpha2
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: input1
+          lower-bound: 0
+        - id: input2
+          lower-bound: 0
+      ports:
+        - id: input_port1
+          type: flow
+        - id: input_port2
+          type: flow
+        - id: output_port
+          type: flow
+      port-field-definitions:
+        - port: input_port1
+          field: flow
+          definition: -input1
+        - port: input_port2
+          field: flow
+          definition: -input2
+        - port: output_port
+          field: flow
+          definition: input1 * alpha1 + input2 * alpha2
+
+    - id: decompose_1_flow_into_2_flow
+      description: A Consumption aggregation model
+      variables:
+        - id: input1
+        - id: input2
+      ports:
+        - id: input_port1
+          type: flow
+        - id: input_port2
+          type: flow
+        - id: output_port
+          type: flow
+      port-field-definitions:
+        - port: input_port1
+          field: flow
+          definition: input1
+        - port: input_port2
+          field: flow
+          definition: input2
+      binding-constraints:
+        - id: Conversion
+          expression: sum_connections(output_port.flow) = input1 + input2
+
+    - id: convertor_receive_in
+      description: A convertor model
+      parameters:
+        - id: alpha
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: input
+          lower-bound: 0
+      ports:
+        - id: input_port
+          type: flow
+        - id: output_port
+          type: flow
+      port-field-definitions:
+        - port: output_port
+          field: flow
+          definition: input * alpha
+      binding-constraints:
+        - id: Conversion
+          expression: sum_connections(input_port.flow) = input
+
+    - id: node_mod
+      description: A node model with two ports
+      variables:
+        - id: p
+      ports:
+        - id: injection_port_n
+          type: flow
+        - id: injection_port_e
+          type: flow
+      port-field-definitions:
+        - port: injection_port_e
+          field: flow
+          definition: p
+      binding-constraints:
+        - id: balance
+          expression:  sum_connections(injection_port_n.flow) = p
+
+    - id: generator_with_co2
+      description: generator model that emits CO2
+      parameters:
+        - id: pmin
+          time-dependent: false
+          scenario-dependent: false
+        - id: pmax
+          time-dependent: false
+          scenario-dependent: false
+        - id: cost
+          time-dependent: false
+          scenario-dependent: false
+        - id: emission_rate
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: p
+          lower-bound: pmin
+          upper-bound: pmax
+      ports:
+        - id: injection_port
+          type: flow
+        - id: co2_port
+          type: emission
+      port-field-definitions:
+        - port: injection_port
+          field: flow
+          definition: p
+        - port: co2_port
+          field: emission
+          definition: p * emission_rate
+      objective: expec(sum(cost * p))
+
+    - id: quota_co2
+      description: A CO2 quota model
+      parameters:
+        - id: quota
+          time-dependent: false
+          scenario-dependent: false
+      ports:
+        - id: emission_port
+          type: emission
+      binding-constraints:
+        - id: bound_co2
+          expression:  sum_connections(emission_port.emission) <= quota
+
+    - id: link
+      description: A link model
+      parameters:
+        - id: f_max
+          time-dependent: false
+          scenario-dependent: false
+      variables:
+        - id: flow
+          lower-bound: -f_max
+          upper-bound: f_max
+      ports:
+        - id: injection_port_from
+          type: flow
+        - id: injection_port_to
+          type: flow
+      port-field-definitions:
+        - port: injection_port_from
+          field: flow
+          definition: -flow
+        - port: injection_port_to
+          field: flow
+          definition: flow
